+++ sdk/eventhubs/azure_messaging_eventhubs/examples/eventhubs_consume_messages.rs
@@ -2,12 +2,12 @@
 // Licensed under the MIT license.
 
 //! This sample demonstrates how to consume events from an Event Hub partition using the `ConsumerClient`.
+use azure_core::time::Duration;
 use azure_identity::DefaultAzureCredential;
 use azure_messaging_eventhubs::{
     ConsumerClient, OpenReceiverOptions, StartLocation, StartPosition,
 };
 use futures::StreamExt;
-use std::time::Duration;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -39,7 +39,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     location: StartLocation::Earliest,
                     ..Default::default()
                 }),
-                receive_timeout: Some(Duration::from_secs(5)),
+                receive_timeout: Some(Duration::seconds(5)),
                 ..Default::default()
             }),
         )