+++ sdk/typespec/typespec_client_core/src/lib.rs
@@ -8,7 +8,6 @@
 mod macros;
 pub mod async_runtime;
 pub mod base64;
-pub mod date;
 pub mod error;
 pub mod fmt;
 pub mod fs;
@@ -18,6 +17,7 @@ pub mod http;
 pub mod json;
 pub mod sleep;
 pub mod stream;
+pub mod time;
 #[cfg(feature = "xml")]
 pub mod xml;